Reply 불러오는 함수 util로 추출
