+++ src/components/pages/reply/ReplyList.tsx
@@ -1,10 +1,8 @@
-import { collection, query, orderBy, onSnapshot } from 'firebase/firestore';
 import React, { useEffect, useState } from 'react';
-
 import { useAuth } from '@/contexts/AuthContext';
 import { Reply } from '@/types/Reply';
 import ReplyRow from './ReplyRow';
-import { firestore } from '../../../firebase';
+import { fetchReplies } from '@/utils/replyUtils';
 
 interface ReplyListProps {
   boardId: string;
@@ -15,19 +13,8 @@ interface ReplyListProps {
 const ReplyList: React.FC<ReplyListProps> = ({ boardId, postId, commentId }) => {
   const [replies, setReplies] = useState<Reply[]>([]);
   const { currentUser } = useAuth();
-
   useEffect(() => {
-    const repliesRef = collection(firestore, `boards/${boardId}/posts/${postId}/comments/${commentId}/replies`);
-    const repliesQuery = query(repliesRef, orderBy('createdAt', 'asc'));
-
-    const unsubscribe = onSnapshot(repliesQuery, (snapshot) => {
-      const fetchedReplies = snapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      })) as Reply[];
-      setReplies(fetchedReplies);
-    });
-
+    const unsubscribe = fetchReplies(boardId, postId, commentId, setReplies);
     return () => unsubscribe();
   }, [boardId, postId, commentId]);
